Use a nonce when signing in
The official firebase docs demonstrate how to sign in with Apple
without using the sign_in_with_apple package so we are not currently
usign this package. Currently we use firebase_auth which has sign in
with Apple functionality built in.

The old docs (https://firebase.flutter.dev/docs/auth/social/#apple) do
use a nonce so we have replicated that code here but this package is
not currently being used.